backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
loguru==0.7.2
sse-starlette

# LangChain and LangGraph for agent orchestration - pinned for compatibility
langgraph>=0.2.45
aiosqlite==0.21.0
langgraph-checkpoint-sqlite
langchain>=0.3.7
langchain-openai>=0.2.11
langchain-core>=0.3.21
langsmith
openai>=1.54.0,<2.0.0
httpx>=0.27.0,<1.0.0

# Additional utilities
python-dotenv
aiosmtplib
schedule
email-validator>=2.0.0  # For GDPR compliance email validation

# Rate limiting and retry logic
slowapi>=0.1.9
tenacity>=8.2.0

# Structured logging
structlog>=24.1.0

# MCP
mcp
